+++ SharePointFramework/PortfolioWebParts/src/loc/mystrings.d.ts
@@ -1,4 +1,6 @@
 declare interface IPortfolioWebPartsStrings {
+  ColumnRenderDescription: ReactNode
+  DisplayNameDescription: string
   HideViewsLabel: string
   ProjectsAccessSearchBoxPlaceholderText: string
   ProjectsAccessHeaderText: string
@@ -19,6 +21,8 @@ declare interface IPortfolioWebPartsStrings {
   ColumnRenderOptionTrend: string
   ColumnRenderOptionTaxonomy: string
   ColumnRenderOptionModal: string
+  ColumnRenderOptionFilenameWithIcon: string
+  ColumnRenderPersistGloballyLabel: string
   DataSourceGroupName: string
   CommandBarGroupName: string
   SearchBoxGroupName: string